Implement basic BAC Calculator with Truncation
The BAC algorythm takes into account a persons weight, gender
and current drink consumption measured in standard drinks. It
uses those values to add to the total BAC level as the drink
is first added. This patch does not take into account absorption
or metabolism of alcohol as the drink is first added.