Implement timer. Fix various BAC issues.
This patch implements the formatted timer. The function generateBACtimer()
will generate the current time left for the person to sober up. This will
need to be called periodically in order to update the time on the display.

This patch fixes the drink counter. Any time the current drink value
is set to something other than zero it needs to be set back to zero or
it will break the refresh button functionality. The current drink value
reflects the drink being consumed in standard drinks. In later patches
we should consider allowing the user to edit this variable manually to
add custom drinks.